Working on alembic migration with batched modifications compatible with sqlite
