+++ alembic/README
@@ -12,12 +12,17 @@ When submitting a pull request it will be helpful to indicate whether any migrat
 
 To create new revisions automatically start with the old version of the database in BBN.sqlite, then run alembic in the new version:
 
-
-git checkout develop
-rm BBN.sqlite
-python alembic/fill_db.py # generate db with old schema
-git checkout master 
-alembic stamp head # Mark version for comparison (set the head)
 git checkout my_branch
 alembic revision --autogenerate -m "brief notes on update" # Create the diff
 git add alembic/versions/abcabc123123_brief_notes_on_update.py # add to version control
+
+To apply: alembic upgrade head
+
+To create the database in the first place:
+python alembic/fill_db.py # generate example db
+
+If you need to manually set the head:
+git checkout master # e.g.
+alembic stamp head # Mark version for comparison (set the head)
+
+You may need to modify the alembic version table in the database manually to get this to work...
