--- alembic/versions/2de6acc8d538_shift_dsp_properties_to_streamselectors.py
@@ -1,59 +1,0 @@
-"""Shift DSP properties to StreamSelectors
-
-Revision ID: 2de6acc8d538
-Revises: 38cb56d154af
-Create Date: 2019-04-02 11:51:43.819235
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '2de6acc8d538'
-down_revision = '38cb56d154af'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('streamselect',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('stream_type', sa.String(), nullable=False),
-    sa.Column('dsp_channel', sa.Integer(), nullable=True),
-    sa.Column('if_freq', sa.Float(), nullable=False),
-    sa.Column('kernel_data', sa.LargeBinary(), nullable=True),
-    sa.Column('kernel_bias', sa.Float(), nullable=False),
-    sa.Column('threshold', sa.Float(), nullable=False),
-    sa.Column('threshold_invert', sa.Boolean(), nullable=False),
-    sa.ForeignKeyConstraint(['id'], ['nodeproxy.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.drop_table('qubitproxy')
-    op.drop_column('receiverchannel', 'stream_type')
-    op.drop_column('receiverchannel', 'kernel_bias')
-    op.drop_column('receiverchannel', 'threshold')
-    op.drop_column('receiverchannel', 'dsp_channel')
-    op.drop_column('receiverchannel', 'if_freq')
-    op.drop_column('receiverchannel', 'kernel_data')
-    op.drop_column('receiverchannel', 'threshold_invert')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('receiverchannel', sa.Column('threshold_invert', sa.BOOLEAN(), nullable=False))
-    op.add_column('receiverchannel', sa.Column('kernel_data', sa.BLOB(), nullable=True))
-    op.add_column('receiverchannel', sa.Column('if_freq', sa.FLOAT(), nullable=False))
-    op.add_column('receiverchannel', sa.Column('dsp_channel', sa.INTEGER(), nullable=True))
-    op.add_column('receiverchannel', sa.Column('threshold', sa.FLOAT(), nullable=False))
-    op.add_column('receiverchannel', sa.Column('kernel_bias', sa.FLOAT(), nullable=False))
-    op.add_column('receiverchannel', sa.Column('stream_type', sa.VARCHAR(), nullable=False))
-    op.create_table('qubitproxy',
-    sa.Column('id', sa.INTEGER(), nullable=False),
-    sa.ForeignKeyConstraint(['id'], ['nodeproxy.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.drop_table('streamselect')
-    # ### end Alembic commands ###
