--- alembic/versions/38cb56d154af_added_calibration_tables.py
@@ -1,44 +1,0 @@
-"""Added Calibration Tables
-
-Revision ID: 38cb56d154af
-Revises: 
-Create Date: 2019-03-22 16:06:55.175177
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '38cb56d154af'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('sample',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('calibration',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('category', sa.String(), nullable=True),
-    sa.Column('value', sa.Float(), nullable=False),
-    sa.Column('uncertainty', sa.Float(), nullable=True),
-    sa.Column('date', sa.DateTime(), nullable=True),
-    sa.Column('sample_id', sa.Integer(), nullable=True),
-    sa.ForeignKeyConstraint(['sample_id'], ['sample.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('calibration')
-    op.drop_table('sample')
-    # ### end Alembic commands ###
