--- alembic/versions/db5d5e5b6830_hash_for_pipeline.py
@@ -1,30 +1,0 @@
-"""Hash for pipeline
-
-Revision ID: db5d5e5b6830
-Revises: 2de6acc8d538
-Create Date: 2019-04-10 10:53:45.411243
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'db5d5e5b6830'
-down_revision = '2de6acc8d538'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('nodeproxy', sa.Column('hash_val', sa.Integer(), nullable=False, server_default='12345'))
-    query = 'update nodeproxy set hash_val=abs(random());'
-    op.execute(query) 
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('nodeproxy', 'hash_val')
-    # ### end Alembic commands ###
