+++ alembic/versions/e16b0aa795e2_2019_2_updates.py
@@ -1,0 +1,43 @@
+"""2019.2 updates
+
+Revision ID: e16b0aa795e2
+Revises: 
+Create Date: 2019-10-28 16:19:21.109468
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = 'e16b0aa795e2'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('dcsource', sa.Column('qubit_id', sa.Integer(), nullable=True))
+    with op.batch_alter_table('dcsource') as batch_op:
+        batch_op.create_foreign_key("fk_dcsource_qubit_id_qubit", 'qubit', ['qubit_id'], ['id'])
+    op.add_column('edge', sa.Column('cnot_impl', sa.String(), nullable=True))
+    op.add_column('qubit', sa.Column('bias_pairs', sa.PickleType(), nullable=True))
+    with op.batch_alter_table('transmitter') as batch_op:
+        batch_op.drop_column('delay')
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('transmitter', sa.Column('delay', sa.FLOAT(), nullable=False))
+    with op.batch_alter_table('qubit') as batch_op:
+        batch_op.drop_column('bias_pairs')
+    with op.batch_alter_table('edge') as batch_op:
+        batch_op.drop_column('cnot_impl')
+    with op.batch_alter_table('dcsource') as batch_op:
+        batch_op.drop_constraint("fk_dcsource_qubit_id_qubit", type_='foreignkey')
+    with op.batch_alter_table('dcsource') as batch_op:
+        batch_op.drop_column('qubit_id')
+    print("I'm here")
+    # ### end Alembic commands ###
